Working on concurrency issues with integration test

+++ grails-app/services/allison/zipcode/ZipcodeService.groovy
@@ -1,7 +1,6 @@
 package allison.zipcode
 
 
-import jsr166y.ForkJoinPool
 import groovyx.gpars.GParsPool
 
 
@@ -25,76 +24,41 @@ class ZipcodeService {
         def dir = DownloadService.getCountryDir(country)
         FileUtils.deleteDirectory(new File(dir))
 
-//        def file
-//        def address
-//        def xml
-//        def allCodes
-//        def zipcode
-//        def slurper = new XmlSlurper()
-
-
-
         def start = System.currentTimeMillis()
 
-//        country.states.each { state ->
-//            def file = DownloadService.getStateFileName(state)
-//            def address = downloadService.getAddress(state)
-//            println "Downloading file: ${file}"
-//            try {
-//
-//                // Download zip codes
-//                DownloadService.download(file, address)
-//
-//                // Only clear the zip codes on a successful download
-//                clearZipcodes(state)
-//            } catch (FileNotFoundException ex) {
-//                throw new UnableToDownloadException(message: "Unable to create ${file} from download")
-//            }
-//
-//        }
         GParsPool.withPool {
-            country.states.eachParallel { state ->
+            country.states.eachParallel { State state ->
                 def file = DownloadService.getStateFileName(state)
-                def address = downloadService.getAddress(state)
-                println "Downloading file: ${file}"
+
                 try {
+                    def address = downloadService.getAddress(state)
+                    println "Downloading file: ${file}"
 
                     // Download zip codes
                     DownloadService.download(file, address)
                     Zipcode.withTransaction {
                         State.withTransaction {
                             // Only clear the zip codes on a successful download
-                            clearZipcodes(state)
+                            ZipcodeService.clearZipcodesThreaded(state)
 
                             // Slurp and save in Domain
                             def xml = new XmlSlurper().parse(file)
-        //                    xml = slurper.parse(file)
-        //                    def allCodes = xml.code
-
                             def zipcode
-                            //for (code in allCodes) {
                             xml.code.each {
                                 zipcode = ZipcodeService.slurpZipcode(it)
                                 ZipcodeService.addZipcodeToState(state, zipcode)
                             }
-
+                            state = State.lock(state.id)
+                            println "Num zipcodes: " + state?.zipcodes?.size()//state?.zipcodes?.size()
                         } // State.withTransaction
                     } // Zipcode.withTransaction
                 } catch (FileNotFoundException ex) {
                     throw new UnableToDownloadException(message: "Unable to create ${file} from download")
+                } catch (UnableToDownloadException ex) {
+                    throw ex
                 }
-                println "Num zipcodes: " + state?.zipcodes?.size()
             } // country.states.each
         } // withPool
-//        session.merge(country)
-//        }// withTransaction
-
-//        // Wait for all the threads to finish
-//        threads.each {
-//            try {
-//                it.join()
-//            } catch (InterruptedException ignore) { println "Interrupted Exception"}
-//        }
 
         println "Total time to load zipcodes: " + (System.currentTimeMillis() - start)
 
@@ -125,39 +89,31 @@ class ZipcodeService {
 
 
     static addZipcodeToState(State state, Zipcode zipcode) {
-        if (state) {
-            // Only add the zipcode if it is valid
-//            state = state.lock(state.id)
-//            state.addToZipcodes(zipcode)
-//            Zipcode.withTransaction {
-//                State.withTransaction {
-                    state = State.lock(state.id) // Without this lock, zipcode validation doesn't work
-                    state.addToZipcodes(zipcode)
-//                    if (zipcode.validate()) {
-//                        state = state.lock(state.id)
-//                        state.save()
-//                    } else {
-//                        state.removeFromZipcodes(zipcode)
-//                        zipcode.discard()
-//                    }
-
-                    zipcode.validate()
-//                    state = State.lock(state.id)
-                    if (zipcode.hasErrors() ||
-                            ((state = State.lock(state.id)) && !state.save(failOnError: true, flush: true))) {
+        State.withTransaction {
+        if (state?.id) {  // Only add to a state that exists
+            println "add zipcode to state: " + state
+            state = State.lock(state.id) // Without this lock, zipcode validation doesn't work
+            println "after lock: " + state
+            state.addToZipcodes(zipcode)
+            state = State.lock(state.id)
+                    if (zipcode.validate()) {
+                        state.save()
+                    } else {
                         state = State.lock(state.id)
                         state.removeFromZipcodes(zipcode)
                         zipcode.discard()
                     }
 
-
-//                    state = state.lock(state.id)
-//                    if (! state.save()) {
-//                        state.removeFromZipcodes(zipcode)
-//                        zipcode.discard()
-//                    }
-//                } // State.withTransaction
-//            } // Zipcode.withTransaction
+//            zipcode.validate()
+////                    state = State.lock(state.id)
+//            if (zipcode.hasErrors() ||
+//                    ((state = State.lock(state.id)) && // Lock before the save
+//                        !state.save(failOnError: true, flush: true))) {
+//                state = State.lock(state.id)
+//                state.removeFromZipcodes(zipcode)
+//                zipcode.discard()
+//            }
+        }
         }
     }
 
@@ -182,6 +138,18 @@ class ZipcodeService {
     }
 
     static clearZipcodes(State state) {
+        def tmp = []
+        if (state?.zipcodes) {
+            tmp.addAll(state.zipcodes)
+            tmp.each { zipcode ->
+                state.removeFromZipcodes(zipcode)
+                zipcode.delete()
+                state.save(flush: true)
+            }
+        }
+    }
+
+    static clearZipcodesThreaded(State state) {
 
 //        if (!country || !state) { // If the country doesn't exist, there is nothing to do
 //            return
@@ -192,12 +160,18 @@ class ZipcodeService {
 //        state = State.lock(state.id)
 //        state?.zipcodes?.each {
             def tmp = []
-            tmp.addAll(state?.zipcodes)
-            tmp.each { zipcode ->
+            state = State.lock(state.id)
+            if (state?.zipcodes) {
                 state = State.lock(state.id)
-                state.removeFromZipcodes(zipcode)
-                zipcode = Zipcode.lock(zipcode.id)
-                zipcode.delete()
+                tmp.addAll(state.zipcodes)
+                tmp.each { zipcode ->
+                    state = State.lock(state.id)
+                    state.removeFromZipcodes(zipcode)
+                    zipcode = Zipcode.lock(zipcode.id)
+                    zipcode.delete()
+                    state = State.lock(state.id)
+                    state.save(flush: true)
+                }
             }
 //        } // Nothing to do if there are no zipcodes
     }
